main file kleiner gemaakt en wat dingen aangepast voor makkelijke leesbaarheid
