+++ RollerCoaster/model/rollerCoasterModel.mtl
@@ -1,12 +1,12 @@
 # Blender 4.4.3 MTL File: 'rollerCoasterModel.blend'
 # www.blender.org
 
-newmtl Material
+newmtl color
 Ns 250.000000
 Ka 1.000000 1.000000 1.000000
 Kd 0.800000 0.800000 0.800000
 Ks 0.500000 0.500000 0.500000
 Ke 0.000000 0.000000 0.000000
-Ni 1.450000
+Ni 1.500000
 d 1.000000
 illum 2
